# Benchmarks and improvements to route value dictionary