@rynowak 2018-06-25
**Before**

|        Method |      Mean |    Error |   StdDev |         Op/s |  Gen 0 | Allocated |
|-------------- |----------:|---------:|---------:|-------------:|-------:|----------:|
| AddSingleItem |  46.63 ns | 1.045 ns | 2.014 ns | 21,444,616.2 | 0.0017 |     144 B |
| AddThreeItems | 109.41 ns | 1.319 ns | 1.234 ns |  9,139,647.0 | 0.0014 |     144 B |
| SetSingleItem |  48.09 ns | 1.206 ns | 1.651 ns | 20,793,881.5 | 0.0016 |     144 B |
| SetThreeItems | 124.91 ns | 2.014 ns | 1.785 ns |  8,006,012.8 | 0.0017 |     144 B |

**After**

|        Method |     Mean |     Error |    StdDev |         Op/s |  Gen 0 | Allocated |
|-------------- |---------:|----------:|----------:|-------------:|-------:|----------:|
| AddSingleItem | 31.69 ns | 0.7823 ns | 2.2066 ns | 31,553,826.8 | 0.0014 |     128 B |
| AddThreeItems | 51.84 ns | 1.2012 ns | 0.8685 ns | 19,289,741.1 | 0.0013 |     128 B |
| SetSingleItem | 29.80 ns | 0.6425 ns | 0.7399 ns | 33,560,686.7 | 0.0014 |     128 B |
| SetThreeItems | 53.40 ns | 1.1301 ns | 1.7595 ns | 18,725,699.7 | 0.0013 |     128 B |