@rynowak 2018-06-26
Here's the results from the Github benchmark

**Before**

|           Method |        Mean |      Error |     StdDev |        Op/s | Scaled | ScaledSD |  Gen 0 | Allocated |
|----------------- |------------:|-----------:|-----------:|------------:|-------:|---------:|-------:|----------:|
|         Baseline |    134.0 ns |   2.504 ns |   2.459 ns | 7,461,686.7 |   1.00 |     0.00 | 0.0305 |   3.63 KB |
|              Dfa |    411.7 ns |   7.808 ns |   7.303 ns | 2,428,915.4 |   3.07 |     0.08 | 0.4883 |  46.31 KB |
|      Instruction |    540.5 ns |   5.882 ns |   5.502 ns | 1,850,172.5 |   4.03 |     0.08 | 0.8545 |  74.54 KB |
|      LegacyRoute | 10,319.9 ns | 205.498 ns | 359.913 ns |    96,900.4 |  77.03 |     2.97 |      - |  53.58 KB |
| LegacyTreeRouter |  1,728.1 ns |  11.000 ns |   9.185 ns |   578,661.6 |  12.90 |     0.23 | 0.4883 |   80.6 KB |

**After**

|           Method |       Mean |      Error |     StdDev |        Op/s | Scaled | ScaledSD |  Gen 0 | Allocated |
|----------------- |-----------:|-----------:|-----------:|------------:|-------:|---------:|-------:|----------:|
|         Baseline |   135.9 ns |   1.323 ns |   1.173 ns | 7,357,398.0 |   1.00 |     0.00 | 0.0610 |   6.05 KB |
|              Dfa |   358.3 ns |   3.300 ns |   2.756 ns | 2,790,727.1 |   2.64 |     0.03 | 0.4883 |  44.83 KB |
|      Instruction |   469.0 ns |   6.916 ns |   5.399 ns | 2,132,249.6 |   3.45 |     0.05 | 0.7324 |  72.54 KB |
|      LegacyRoute | 9,813.1 ns | 156.370 ns | 138.618 ns |   101,904.5 |  72.20 |     1.15 |      - |  52.09 KB |
| LegacyTreeRouter | 1,647.6 ns |  17.063 ns |  15.961 ns |   606,955.3 |  12.12 |     0.15 | 0.4883 |  79.12 KB |