@rynowak 2018-06-26
Updated this. @kichalla - it would be great if you could take a quick look.